backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pymongo>=4.6.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
